[fix] Detaching entity after save. (#444)
Trying to fix `The instance of entity type 'EventInfo' cannot be tracked because another instance with the same key value for {'EventInfoId'} is already being tracked. When attaching existing entities, ensure that only one entity instance with a given key value is attached. Consider using 'DbContextOptionsBuilder.EnableSensitiveDataLogging' to see the conflicting key values.``